+++ test/CommandLine/CommandLineTest.cs
@@ -1,14 +1,10 @@
-﻿using System;
-using System.Text;
-using System.Collections.Generic;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace RJCP.Core.CommandLine
+﻿namespace RJCP.Core.CommandLine
 {
-    /// <summary>
-    /// Summary description for UnitTest1
-    /// </summary>
-    [TestClass]
+    using System;
+    using System.Collections.Generic;
+    using NUnit.Framework;
+
+    [TestFixture]
     public class CommandLineTest
     {
         private class NoArguments
@@ -157,48 +153,48 @@ namespace RJCP.Core.CommandLine
             public IList<string> Arguments { get { return m_Arguments; } }
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(ArgumentNullException))]
         public void CmdLine_NullOptions()
         {
             Options options = Options.Parse(null, null);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_NullArguments()
         {
             OptionalArguments myOptions = new OptionalArguments();
             Options options = Options.Parse(myOptions, null);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_NoArguments()
         {
             OptionalArguments myOptions = new OptionalArguments();
             Options options = Options.Parse(myOptions, new string[0], OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_NullArgumentsEmptyOptions()
         {
             NoArguments myOptions = new NoArguments();
             Options options = Options.Parse(myOptions, null, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_NoArgumentsEmptyOptions()
         {
             NoArguments myOptions = new NoArguments();
             Options options = Options.Parse(myOptions, new string[0], OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_SingleShortOption()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -209,8 +205,8 @@ namespace RJCP.Core.CommandLine
             Assert.IsNull(myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsJoined()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -221,8 +217,8 @@ namespace RJCP.Core.CommandLine
             Assert.IsNull(myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsSeparated()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -233,8 +229,8 @@ namespace RJCP.Core.CommandLine
             Assert.IsNull(myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsOnProperties()
         {
             PropertyOptions myOptions = new PropertyOptions();
@@ -245,8 +241,8 @@ namespace RJCP.Core.CommandLine
             Assert.IsNull(myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsStringOneArg()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -257,8 +253,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("foo", myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsStringTwoArgs()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -269,8 +265,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("foo", myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsStringOneArgAssigned()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -281,8 +277,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("foo", myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsStringOneArgJoined()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -293,8 +289,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("foo", myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsStringTwoArgsJoined1()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -305,8 +301,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("foo", myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsStringTwoArgsJoined2()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -319,8 +315,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("foo", options.Arguments[0]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsRequired()
         {
             RequiredArguments myOptions = new RequiredArguments();
@@ -331,8 +327,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("string", myOptions.SearchString);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsRequiredOneArgument1()
         {
             RequiredArguments myOptions = new RequiredArguments();
@@ -343,8 +339,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("string", myOptions.SearchString);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsRequiredOneArgument2()
         {
             RequiredArguments myOptions = new RequiredArguments();
@@ -355,8 +351,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("string", myOptions.SearchString);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ShortOptionsRequiredOneArgument3()
         {
             RequiredArguments myOptions = new RequiredArguments();
@@ -367,8 +363,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("string", myOptions.SearchString);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionMissingArgumentException))]
         public void CmdLineUnix_ShortOptionsRequiredMissingArgument()
         {
@@ -376,8 +372,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "-ifs" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof (OptionMissingException))]
         public void CmdLineUnix_MissingOption()
         {
@@ -385,8 +381,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "-ab" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_RequiredOption()
         {
             RequiredOptions myOptions = new RequiredOptions();
@@ -395,8 +391,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("bar", myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_LongOptionOneArgument()
         {
             RequiredArguments myOptions = new RequiredArguments();
@@ -407,8 +403,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("string", myOptions.SearchString);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_LongOptionTwoArguments()
         {
             RequiredArguments myOptions = new RequiredArguments();
@@ -419,8 +415,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("string", myOptions.SearchString);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionMissingArgumentException))]
         public void CmdLineUnix_LongOptionsRequiredMissingArgument()
         {
@@ -428,8 +424,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "--search" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_LongOptionsBoolean()
         {
             RequiredArguments myOptions = new RequiredArguments();
@@ -440,8 +436,8 @@ namespace RJCP.Core.CommandLine
             Assert.IsNull(myOptions.SearchString);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void cmdLineUnix_LongOptionOnly()
         {
             OptionalLongArguments myOptions = new OptionalLongArguments();
@@ -452,8 +448,8 @@ namespace RJCP.Core.CommandLine
             Assert.IsNull(myOptions.OptionC);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionUnknownException))]
         public void cmdLineUnix_LongOptionOnlyShort()
         {
@@ -461,8 +457,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "-a" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionUnknownException))]
         public void CmdLineUnix_UnknownOption1()
         {
@@ -470,8 +466,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "--search", "string", "-ab", "-c" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionUnknownException))]
         public void CmdLineUnix_UnknownOption2()
         {
@@ -479,8 +475,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "-dabc" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionUnknownException))]
         public void CmdLineUnix_UnknownLongOption()
         {
@@ -488,8 +484,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "--foobar" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionException))]
         public void CmdLineUnix_ExtraArgument()
         {
@@ -497,8 +493,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "-ab", "argument", "-c" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ListOption()
         {
             ListOptions myOptions = new ListOptions();
@@ -510,8 +506,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("test3", myOptions.List[2]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ListOptionSingleElement()
         {
             ListOptions myOptions = new ListOptions();
@@ -521,8 +517,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("test1", myOptions.List[0]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ListOptionMultipleArguments()
         {
             ListOptions myOptions = new ListOptions();
@@ -534,8 +530,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("test3", myOptions.List[2]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ListOptionQuoted1()
         {
             ListOptions myOptions = new ListOptions();
@@ -547,8 +543,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("test 3", myOptions.List[2]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ListOptionQuoted2()
         {
             ListOptions myOptions = new ListOptions();
@@ -560,8 +556,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("test 3", myOptions.List[2]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionException))]
         public void CmdLineUnix_ListOptionQuotedInvalid1()
         {
@@ -569,8 +565,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "-l", @"test1,'test2,test3b','test 3'x" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ListOptionWindowsPath()
         {
             ListOptions myOptions = new ListOptions();
@@ -579,8 +575,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual(@"c:\users\homeuser", myOptions.List[0]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ListOptionQuotedEndEscape()
         {
             ListOptions myOptions = new ListOptions();
@@ -591,8 +587,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual(@"testx\", myOptions.List[2]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionException))]
         public void CmdLineUnix_ListOptionQuotedInvalid3()
         {
@@ -600,8 +596,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "-l", "test1,test2,\"test3" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_StopParsing()
         {
             OptionalArguments myOptions = new OptionalArguments();
@@ -614,8 +610,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("-c", options.Arguments[0]);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_DefaultValueProvidedShort()
         {
             DefaultValueOption myOptions = new DefaultValueOption();
@@ -626,8 +622,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("3", myOptions.Verbosity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_DefaultValueProvidedShort2()
         {
             DefaultValueOption myOptions = new DefaultValueOption();
@@ -638,8 +634,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("3", myOptions.Verbosity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_DefaultValueShort()
         {
             DefaultValueOption myOptions = new DefaultValueOption();
@@ -650,8 +646,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("0", myOptions.Verbosity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_DefaultValueShortProvidedExistingOption()
         {
             DefaultValueOption myOptions = new DefaultValueOption();
@@ -662,8 +658,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("b", myOptions.Verbosity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_DefaultValueProvidedLongEquals()
         {
             DefaultValueOption myOptions = new DefaultValueOption();
@@ -674,8 +670,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("2", myOptions.Verbosity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_DefaultValueProvidedLong()
         {
             DefaultValueOption myOptions = new DefaultValueOption();
@@ -686,8 +682,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("2", myOptions.Verbosity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_DefaultValueLong()
         {
             DefaultValueOption myOptions = new DefaultValueOption();
@@ -698,8 +694,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("0", myOptions.Verbosity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_DefaultValueLong2()
         {
             DefaultValueOption myOptions = new DefaultValueOption();
@@ -710,8 +706,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual("0", myOptions.Verbosity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_TypesEnum()
         {
             TypesOptions myOptions = new TypesOptions();
@@ -720,8 +716,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual(BasicColor.Yellow, myOptions.Color);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_TypesEnumInt()
         {
             TypesOptions myOptions = new TypesOptions();
@@ -730,8 +726,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual(BasicColor.Cyan, myOptions.Color);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_TypesInt()
         {
             TypesOptions myOptions = new TypesOptions();
@@ -740,8 +736,8 @@ namespace RJCP.Core.CommandLine
             Assert.AreEqual(100, myOptions.Opacity);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         [ExpectedException(typeof(OptionFormatException))]
         public void CmdLineUnix_TypesIntInvalid()
         {
@@ -749,8 +745,8 @@ namespace RJCP.Core.CommandLine
             Options options = Options.Parse(myOptions, new[] { "-O", "xxx" }, OptionsStyle.Unix);
         }
 
-        [TestMethod]
-        [TestCategory("CommandLine")]
+        [Test]
+        [Category("Utilities.CommandLine")]
         public void CmdLineUnix_ArgumentsAttribute()
         {
             ArgumentsAttributeOptions myOptions = new ArgumentsAttributeOptions();