+++ docs/HISTORY.md
@@ -10,13 +10,13 @@ in other of my projects (starting with Subversion, then later Git). This reposit
 maintains source history, with the following changes:
 
 * The solution and project files aren't original. The original project started
-  with Visual Studio 2010, but this repository starts with Visual Studio 2015
-  (which also supports MSTest for .NET 4.0).
+  with Visual Studio 2010, and now can be used with Visual Studio 2019.
 * Namespaces have been rewritten to be `RJCP.Core`, as the other projects were
   similar code, but different namespaces while I developed the code for my
   personal use and used it in side projects where I work.
 * The directory structure is different than the original, so it is easier to
   track history.
+* NUnit 2.x is imported as a NuGet package.
 
 The original project started with .NET 4.0. It continues to this time to still
 target .NET 4.0.